refactor: moved out loaders
